# src/main/java/org/barbershop/employee/adapter/in/rest/EmployeeRequestDTO.java
package org.barbershop.employee.adapter.in.rest;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class EmployeeRequestDTO {
    private String name;
    private String role;
    private String phone;
    private String email;
    private Boolean active;
}