# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Barber Shop API
  version: 1.0.0
  description: API para el catálogo, gestión de barbería y ventas.
servers:
  - url: /api
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [Employees]
      summary: Listar empleados con paginación
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
          description: Buscar por nombre
        - name: role
          in: query
          required: false
          schema: { type: string, enum: [BARBER, RECEPTIONIST, MANAGER] }
          description: Filtrar por rol
        - name: active
          in: query
          required: false
          schema: { type: boolean }
          description: Filtrar por estado activo/inactivo
        - name: page
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
          description: Número de página (1-indexed)
        - name: pageSize
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 100 }
          description: Cantidad de items por página (máximo 100)
      responses:
        '200':
          description: Lista de empleados con paginación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedEmployees' }
    post:
      operationId: createEmployee
      tags: [Employees]
      summary: Crear un nuevo empleado
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/EmployeeRequest' } }
      responses:
        '201':
          description: Empleado creado
          content:
            application/json: { schema: { $ref: '#/components/schemas/Employee' } }
        '400':
          $ref: '#/components/responses/BadRequest'
  /employees/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: getEmployee
      tags: [Employees]
      summary: Obtener un empleado por ID
      responses:
        '200': { description: Empleado, content: { application/json: { schema: { $ref: '#/components/schemas/Employee' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateEmployee
      tags: [Employees]
      summary: Actualizar un empleado
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/EmployeeRequest' } } }
      responses:
        '200': { description: Empleado actualizado, content: { application/json: { schema: { $ref: '#/components/schemas/Employee' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      operationId: deleteEmployee
      tags: [Employees]
      summary: Eliminar un empleado
      responses:
        '204': { description: Empleado eliminado }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /sales:
    get:
      operationId: listSales
      tags: [Sales]
      summary: Listar ventas con paginación
      parameters:
        - name: page
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
          description: Número de página (1-indexed)
        - name: pageSize
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 100 }
          description: Cantidad de items por página (máximo 100)
      responses:
        '200':
          description: Lista de ventas con paginación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedSales' }
    post:
      operationId: createSale
      tags: [Sales]
      summary: Crear una nueva venta con items
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SaleRequest' } } }
      responses:
        '201': { description: Venta creada, content: { application/json: { schema: { $ref: '#/components/schemas/Sale' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /sales/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: getSale
      tags: [Sales]
      summary: Obtener una venta por ID
      responses:
        '200': { description: Venta, content: { application/json: { schema: { $ref: '#/components/schemas/Sale' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /reports/daily/{date}:
    parameters:
      - { name: date, in: path, required: true, schema: { type: string, format: date } }
    get:
      operationId: getDailyReport
      tags: [Reports]
      summary: Obtener reporte diario
      responses:
        '200':
          description: Reporte diario
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DailyReport' }
        '404': { $ref: '#/components/responses/NotFound' }
  /reports/daily/generate/{date}:
    parameters:
      - { name: date, in: path, required: true, schema: { type: string, format: date } }
    post:
      operationId: generateDailyReport
      tags: [Reports]
      summary: Generar reporte diario
      responses:
        '201':
          description: Reporte generado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DailyReport' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  responses:
    NotFound:
      description: Recurso no encontrado
    BadRequest:
      description: Solicitud inválida
  schemas:
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages, hasNextPage]
      properties:
        page:
          type: integer
          description: Número de página actual
        pageSize:
          type: integer
          description: Items por página
        total:
          type: integer
          description: Total de items
        totalPages:
          type: integer
          description: Total de páginas
        hasNextPage:
          type: boolean
          description: Si existe siguiente página
    
    # Items
...
    PaginatedEmployees:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Employee' }
        pagination:
          $ref: '#/components/schemas/Pagination'
    Employee:
      allOf:
        - $ref: '#/components/schemas/EmployeeRequest'
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: integer, format: int64, readOnly: true }
            createdAt: { type: string, format: date-time, readOnly: true }
    EmployeeRequest:
      type: object
      required: [name, role]
      properties:
        name: { type: string, maxLength: 100, minLength: 1 }
        role: { type: string, enum: [BARBER, RECEPTIONIST, MANAGER] }
        phone: { type: string, maxLength: 20 }
        email: { type: string, maxLength: 100, format: email }
        active: { type: boolean, default: true }
    
    # Appointments
...
    PaginatedSales:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Sale' }
        pagination:
          $ref: '#/components/schemas/Pagination'
    Sale:
      type: object
      required: [id, customerId, employeeId, paymentMethod, totalAmount, discount, soldAt]
      properties:
        id: { type: integer, format: int64, readOnly: true }
        customerId: { type: integer, format: int64 }
        employeeId: { type: integer, format: int64 }
        paymentMethod: { type: string, enum: [CASH, TRANSFER, CARD] }
        totalAmount: { type: number, format: double, readOnly: true }
        discount: { type: number, format: double, default: 0 }
        notes: { type: string, maxLength: 500 }
        items:
          type: array
          items: { $ref: '#/components/schemas/SaleItemDetail' }
        soldAt: { type: string, format: date-time, readOnly: true }
    SaleRequest:
      type: object
      required: [customerId, employeeId, paymentMethod, items]
      properties:
        customerId: { type: integer, format: int64 }
        employeeId: { type: integer, format: int64 }
        paymentMethod: { type: string, enum: [CASH, TRANSFER, CARD] }
        discount: { type: number, format: double, default: 0 }
        notes: { type: string, maxLength: 500 }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SaleItemRequest' }
    SaleItemRequest:
      type: object
      required: [itemId, quantity, unitPrice]
      properties:
        itemId: { type: integer, format: int64 }
        quantity: { type: integer, minimum: 1 }
        unitPrice: { type: number, format: double, minimum: 0 }
    SaleItemDetail:
      allOf:
        - $ref: '#/components/schemas/SaleItemRequest'
        - type: object
          required: [id, saleId]
          properties:
            id: { type: integer, format: int64, readOnly: true }
            saleId: { type: integer, format: int64, readOnly: true }
    
    # Reports
    DailyReport:
      type: object
      required: [id, reportDate, totalSales, totalTransactions]
      properties:
        id: { type: integer, format: int64, readOnly: true }
        reportDate: { type: string, format: date }
        totalSales: { type: number, format: double }
        totalTransactions: { type: integer }
        cashSales: { type: number, format: double }
        cardSales: { type: number, format: double }
        transferSales: { type: number, format: double }
        servicesSales: { type: number, format: double }
        productsSales: { type: number, format: double }
        generatedAt: { type: string, format: date-time, readOnly: true }
    
    # Audit Logs
...